Desktop/crypto_tracker/requirements.txt
coinbase-python
kafka-python
pandas
influxdb-client
python-dotenv
requests
flask